Added the changes of mass of Jupiter

+++ Program/Program/models.hpp
@@ -10,10 +10,10 @@
 
 #include <string>
 
-void es_euler(const int years, const std::string folder);
-void ejs_verlet(const int years, const std::string folder);
-void ejs_rmc_euler(const int years, const std::string folder);
-void ejs_rmc_verlet(const int years, const std::string folder);
-void fs_rmc_verlet(const int years, const std::string folder);
-void free_model(const int years, const std::string folder);
-void escape(const int years, std::string folder);
+void es_euler(const double years, const std::string folder);
+void ejs_verlet(const double years, const std::string folder);
+void ejs_rmc_euler(const double years, const std::string folder);
+void ejs_rmc_verlet(const double years, const std::string folder);
+void fs_rmc_verlet(const double years, const std::string folder);
+void free_model(const double years, const std::string folder);
+void escape(const double years, std::string folder);
